Moved AudioKit Testing functions to their own file
